requirements.txt
aiohttp==3.9.3
aiolimiter==1.1.0
appdirs==1.4.4
biomart==0.9.2
cutadapt==4.6
defusedxml==0.7.1
graphviz==0.20.1
grid_strategy==0.0.1
hdbscan==0.8.33
joblib==1.3.2
matplotlib==3.8.2
matplotlib_venn==0.11.10
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
pairwisedist==1.3.1
pandas==2.2.0
pydeseq2==0.4.4
PyQt5==5.15.10
PyQt5_sip==12.13.0
pyvis==0.3.2
PyYAML==6.0.1
PyYAML==6.0.1
qdarkstyle==3.2.3
#qpageview==0.6.2
Requests==2.31.0
rnanorm==2.1.0
sanbomics==0.1.0
scanpy==1.9.8
scikit_learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
statsmodels==0.14.1
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.9.0
ui==0.1.4
upsetplot==0.9.0
xlmhglite==1.1.0
